view property on state (resolves circular import)
