--- openhands/memory/__init__.py
@@ -1,3 +1,0 @@
-from openhands.memory.condenser import Condenser
-
-__all__ = ['Condenser']
